+++ src/models/__init__.py
@@ -4,4 +4,5 @@ from sqlalchemy.orm import DeclarativeBase
 class Base(DeclarativeBase):
     pass
 
-from src.models.clubs import Club, ClubMember
+
+from .clubs import ClubMember, Club
